разнес bot_utils на модули поменьше
