+++ bot_handlers.py
@@ -5,6 +5,8 @@ from aiogram import Dispatcher, types, executor
 from config import *
 import db
 import bot_utils
+import keyboards
+import consts
 import playlist_api
 import music_api
 import core
@@ -18,13 +20,13 @@ async def start_handler(message):
     if message.chat.id < 0:
         return
 
-    await bot.send_message(message.chat.id, bot_utils.TEXT['start'])
-    await bot.send_message(message.chat.id, bot_utils.TEXT['menu'], reply_markup=bot_utils.keyboard_start)
+    await bot.send_message(message.chat.id, consts.TEXT['start'])
+    await bot.send_message(message.chat.id, consts.TEXT['menu'], reply_markup=keyboards.keyboard_start)
 
 
 @dp.message_handler(commands=['cancel'])
 async def cancel(message):
-    await bot.send_message(message.chat.id, bot_utils.TEXT['menu'], reply_markup=bot_utils.keyboard_start)
+    await bot.send_message(message.chat.id, consts.TEXT['menu'], reply_markup=keyboards.keyboard_start)
 
 
 @dp.message_handler(lambda m: m.chat.id == ADMINS_CHAT_ID, commands=['next'])
@@ -99,7 +101,7 @@ async def message_handler(message):
     # Пользователь скинул аудио
     if message.audio and message.chat.id != ADMINS_CHAT_ID:
         return await bot.send_audio(message.chat.id, message.audio.file_id, 'Теперь выбери день',
-                                    reply_markup=bot_utils.keyboard_day())
+                                    reply_markup=keyboards.keyboard_day())
 
     # Форс реплаи
     if message.reply_to_message and message.reply_to_message.from_user.id == (await bot.me).id:
@@ -111,13 +113,13 @@ async def message_handler(message):
                 await core.admin_reply(message)
 
         # Ввод названия песни
-        if message.reply_to_message.text == bot_utils.TEXT['predlozka_choose_song']:
+        if message.reply_to_message.text == consts.TEXT['predlozka_choose_song']:
             await core.search_audio(message)
 
         # Обратная связь
-        if message.reply_to_message.text == bot_utils.TEXT['feedback']:
-            await bot.send_message(message.chat.id, bot_utils.TEXT['feedback_thanks'],
-                                   reply_markup=bot_utils.keyboard_start)
+        if message.reply_to_message.text == consts.TEXT['feedback']:
+            await bot.send_message(message.chat.id, consts.TEXT['feedback_thanks'],
+                                   reply_markup=keyboards.keyboard_start)
             await bot.forward_message(ADMINS_CHAT_ID, message.chat.id, message.message_id)
 
         return
@@ -128,32 +130,32 @@ async def message_handler(message):
     # Кнопки
 
     # Кнопка 'Что играет?'
-    if message.text == bot_utils.btn['what_playing']:
+    if message.text == keyboards.btn['what_playing']:
         playback = await playlist_api.get_now()
         if not playback:
-            await bot.send_message(message.chat.id, "Не знаю(", reply_markup=bot_utils.keyboard_what_playing)
+            await bot.send_message(message.chat.id, "Не знаю(", reply_markup=keyboards.keyboard_what_playing)
         else:
-            await bot.send_message(message.chat.id, bot_utils.TEXT['what_playing'].format(*playback),
-                                   reply_markup=bot_utils.keyboard_what_playing)
+            await bot.send_message(message.chat.id, consts.TEXT['what_playing'].format(*playback),
+                                   reply_markup=keyboards.keyboard_what_playing)
 
     # Кнопка 'Предложить песню'
-    elif message.text == bot_utils.btn['predlozka'] or message.text == '/song':
-        await bot.send_message(message.chat.id, bot_utils.TEXT['predlozka_choose_song'],
+    elif message.text == keyboards.btn['predlozka'] or message.text == '/song':
+        await bot.send_message(message.chat.id, consts.TEXT['predlozka_choose_song'],
                                reply_markup=types.ForceReply())
-        await bot.send_message(message.chat.id, bot_utils.TEXT['predlozka_inline_search'],
-                               reply_markup=bot_utils.keyboard_predlozka_inline)
+        await bot.send_message(message.chat.id, consts.TEXT['predlozka_inline_search'],
+                               reply_markup=keyboards.keyboard_predlozka_inline)
 
     # Кнопка 'Хочу в команду'
-    elif message.text == bot_utils.btn['feedback']:
-        await bot.send_message(message.chat.id, bot_utils.TEXT['feedback'], reply_markup=types.ForceReply())
+    elif message.text == keyboards.btn['feedback']:
+        await bot.send_message(message.chat.id, consts.TEXT['feedback'], reply_markup=types.ForceReply())
 
-    elif message.text == bot_utils.btn['help'] or message.text == '/help':
-        await bot.send_message(message.chat.id, bot_utils.TEXT['help']['first_msg'],
-                               reply_markup=bot_utils.keyboard_help)
+    elif message.text == keyboards.btn['help'] or message.text == '/help':
+        await bot.send_message(message.chat.id, consts.TEXT['help']['first_msg'],
+                               reply_markup=keyboards.keyboard_help)
 
     else:
         await bot.forward_message(ADMINS_CHAT_ID, message.chat.id, message.message_id)
-        await bot.send_message(message.chat.id, bot_utils.TEXT['unknown_cmd'], reply_markup=bot_utils.keyboard_start)
+        await bot.send_message(message.chat.id, consts.TEXT['unknown_cmd'], reply_markup=keyboards.keyboard_start)
 
 
 @dp.inline_handler()
@@ -165,9 +167,9 @@ async def query_text(inline_query):
 async def edited_message(message):
     if message.reply_to_message is None:
         return
-    if message.reply_to_message.text == bot_utils.TEXT['predlozka_choose_song']:
+    if message.reply_to_message.text == consts.TEXT['predlozka_choose_song']:
         await message_handler(message)
 
 
 if __name__ == '__main__':
-    executor.start_polling(dp)
+    executor.start_polling(dp, skip_updates=True)
