+++ consts.py
@@ -1,0 +1,87 @@
+from pathlib import Path
+
+
+TEXT = {
+    'start': '''Привет, это бот РадиоКПИ. 
+Ты можешь:
+ - Заказать песню   
+ - Задать любой вопрос
+ - Узнать, что играет сейчас, играло или будет играть
+ - Узнать, как стать частью ламповой команды РадиоКПИ.
+
+⁉️Советуем первым делом прочитать инструкцию /help
+''',
+    'menu': 'Выбери, что хочешь сделать 😜',
+
+    'help': {
+        'orders': '''
+📝Есть 3 способа <b>заказать песню:</b>
+- Нажать на кнопку <code>Заказать песню</code> и ввести название, бот выберет наиболее вероятный вариант
+- Использовать инлайн режим поиска (ввести <code>@kpiradio_bot</code> или нажать на соответствующую кнопку). 
+    В этом случае ты можешь выбрать из 50 найденных вариантов, с возможностью сначала послушать
+- Загрузить или переслать боту желаемую песню
+После этого необходимо выбрать день и время для заказа.''',
+
+        'criteria': '''
+ <b>❗️Советы:</b>
+- Не отправляйте слишком много песен сразу, их все еще принимают люди, а не нейросети
+- Учтите, что у песен с нехорошими словами или смыслом высокий шанс не пройти модерацию
+- Приветствуются песни на украинском <i>(гугл "квоты на радио")</i>
+- Приветствуются нейтральные, спокойные песни, которые не будут отвлекать от учебного процесса ''',
+
+        'playlist': '''
+<b>⏭Плейлист радио:</b>
+- Узнать что играет сейчас, играло до этого или будет играть можно нажав на кнопку "<code>Что играет</code>"
+- Помните дату и время когда играла песня и хотите ее найти? Можете найти ее там же -> <code>Поиск песни по времени</code>
+- Если вы заказываете песню во время эфира, мы постараемся сделать так, что бы она заиграла следующей
+- Заказ <b>одноразовый</b>. Если ваша песня не успела войти в эфир - закажите <b>на следующий раз</b>.''',
+
+        'feedback': '''
+  <b>🖌Обратная связь:</b>
+- Вы всегда можете написать команде админов что вы думаете о них и о радио
+- Если хотите стать частью радио - пишите об этом и готовьтесь к анкетам
+- Считаете что то неудобным? Есть предложения по улучшению? Не задумываясь пиши нам.      
+        ''',
+
+        'btns': {
+            'orders': '📝Заказ песни',
+            'criteria': '❗Критерии модерации',
+            'playlist': '⏭Тонкости плейлиста',
+            'feedback': '🖌Обратная связь'
+        },
+        'first_msg': 'Выберите интересующую вас тему. (Советуем прочитать все)',
+    },
+
+    'feedback': 'Ты можешь оставить отзыв о работе РадиоКПИ или предложить свою кандидатуру для вступления в наши ряды! \n \
+Напиши сообщение и админы ответят тебе! \n(⛔️/cancel)',
+    'feedback_thanks': 'Спасибо за заявку, мы обязательно рассмотрим ее!',
+
+    'predlozka_choose_song': 'Что ты хочешь услышать? Напиши название или скинь свою песню\n(⛔️/cancel)',
+    'predlozka_inline_search': 'Нажми на кнопку для более удобного поиска 👀',
+
+    'predlozka_moderating': 'Спасибо за заказ ({0}), ожидайте модерации!',  # время
+    'predlozka_ok': 'Ваш заказ ({0}) принят!',  # название песни
+    'predlozka_ok_next': 'Ваш заказ ({0}) принят и заиграет {1}',
+    'predlozka_neok': 'Ваш заказ ({0}) отклонен(',
+
+    'what_playing': """⏮ <b>Предыдущий трек: </b> {0}, 
+▶️ <b>Сейчас играет: </b> {1}
+⏭ <b>Следующий трек: </b> {2}""",
+
+    'song_no_prev': 'Не знаю(',
+    'song_no_next': 'Доступно только во время эфира',
+
+    'error': 'Не получилось(',
+    'unknown_cmd': 'Шо ты хош? Для заказа песни не забывай нажимать на кнопку "Заказать песню". Помощь тут /help',
+
+    'days1': ['Понедельник', 'Вторник', 'Среда', 'Четверг', 'Пятница', 'Суббота', 'Воскресенье'],
+    'days2': ['Сегодня', 'Завтра', 'Послезавтра', 'Сейчас'],
+    'times': ['Утрений эфир', 'Первый перерыв', 'Второй перерыв', 'Третий перерыв', 'Четвертый перерыв',
+              'Вечерний эфир'],
+}
+
+paths = {
+    'orders': Path('D:/Вещание Радио/Заказы'),
+    'archive': Path('D:/Вещание Радио/Архив'),
+    'ether': Path('D:/Вещание Радио/Эфир'),
+}
